desafio4 finalizado

+++ desafio4-1raPreentrega/src/controllers/ProductManager.js
@@ -1,6 +1,6 @@
 import { Console } from 'console';
 import {copyFile, promises as fs} from 'fs';
-//import { nanoid } from 'nanoid';
+import { nanoid } from 'nanoid';
 
 export default class ProductManager {
     constructor(){
@@ -8,47 +8,35 @@ export default class ProductManager {
         this.products = []
     }
 
-    static id=0;
-
     writeProduct = async (product)=>{
         await fs.writeFile(this.path, JSON.stringify(product));
     }
 
-    addProduct = async (product)=>{
-        
-        ProductManager.id++;
-        //console.log(ProductManager.id)    ;   
-        let productsOld = await this.readProducts();
-        product.id= ProductManager.id;
-        let productsAll = [...productsOld,product];
-        await fs.writeFile(this.path, JSON.stringify(productsAll));
-        //await this.writeProduct(productsAll);
-        return "Producto Agregado"
-
-        // this.products.push(newProduct);
-        // await fs.writeFile(this.path,JSON.stringify(this.products))
-    }
-    
     readProducts = async ()=>{
         let products = await fs.readFile(this.path,'utf-8');
-        
         return JSON.parse(products)
     }
 
-
+    addProduct = async (product)=>{
+        let productsOld = await this.readProducts();
+        product.id = nanoid();
+        let productsAll = [...productsOld,product];
+        await this.writeProduct(productsAll);
+        return `Producto ${product.title} agregado correctamente`
+     }
+ 
     getProducts = async ()=>{
         let products = await this.readProducts();
-        //console.log(products)
         return products;
-
     }
     
     exist = async(id) =>{
         let products = await this.readProducts();
+        //console.log(`console exist ${id}`)
+        //console.log(products)
         return products.find(el=>el.id === id)
     }
     getProductById = async(id)=>{
-        
         let product = await this.exist(id);
         if(!product) {
             console.log(`Producto ${id} no existe`)
@@ -58,21 +46,27 @@ export default class ProductManager {
                 return product}
     }
 
-    updateProduct = async ({id, ...product})=> {
-        let idProd = product.id;
+    updateProduct = async (id, product)=> {
+        //console.log(id)
+        //console.log(product)
+        let productToUpdate = await this.exist(id);
+        if(!productToUpdate) {
+            console.log(`Producto ${id} no existe`)
+            return "Producto no existe"    
+        }
         await this.deleteProductByID(id);
-        let productOld = await this.readProducts();
+        let productsAll = await this.readProducts();
+        
         let productsNew = [
-            {id, ...product},
-            ...productOld
-        ]
-        await this.path,JSON.stringify(this.productsNew)
+            {...product, id:id},
+            ...productsAll
+        ];
+        await this.writeProduct(productsNew);
+        return `Producto ${product.title} modificado exitosamente`
     }
 
     deleteProductByID = async (id)=>{
         let products = await this.readProducts();
-        //let newProducts = products.filter(el=>el.id != id)
-        //console.log(newProducts)
         let deleteProduct = products.some(el=>el.id === id);
         if(deleteProduct) {
             let newProducts = products.filter(el=>el.id != id);
